lint update

+++ src/libraries/FileTypes.ts
@@ -65,7 +65,7 @@ class FileTypes {
       // Convert to Uint8Array
       signature = new Uint8Array(signature, 0, 10)
     }
-    const type = types.find(library => library.signature && this.bufferIsEqual(library.signature, signature as Uint8Array))
+    const type = types.find(library => library.signature && this.bufferIsEqual(library.signature, signature))
     return type ? new FileType(type) : undefined
   }
